refactor: extract MixedAudioTypes.swift from MixedAudioRecorder.swift
Extracts 2 support types (MixedAudioSourceStartTimes, MixedAudioTrackInsertionOffsets)
from MixedAudioRecorder.swift into a new MixedAudioTypes.swift.

Byte-identical move; SHA-256 3d686bdb9f4c08d5e68b9cc987b9426b0701b969bbb4311292c817bd0a5be374
verified against origin/main:MixedAudioRecorder.swift lines 328-367.

Imports: CoreMedia (for CMTime) + Foundation.

Reduces MixedAudioRecorder.swift from 381 to 340 lines.

Tests: 667 (unchanged).
